browser class that can execute relevant action steps and produce observations. Separate from workarena example
